TDD working version of RPS
