--- spec/features/game_result_spec.rb
@@ -1,1 +1,0 @@
-
