+++ views/result.erb
@@ -14,11 +14,7 @@
     <div style="width:800px; margin: 0 auto; text-align: center;">
       <h1 style="font-size: 50px;"><%= @player1_name %>, you chose <%= @player1_choice %>!</h1><br><br><br><br>
       <p style="font-size: 20px;">The computer chose ... <br><br> ... <br><br> ... <br><br> ... </p>
-
-
-      <%# insert line here as to what the computer chose %>
-
-
+      <h1 style="font-size: 100px;"><%= @result %></h1>
     </div>
   </body>
